requirements.txt
fastapi==0.128.2
google-auth==2.48.0
jinja2==3.1.6
python-multipart==0.0.22
requests==2.32.5
uvicorn==0.40.0
azure-storage-blob==12.27.1
pymongo==4.16.0